+++ Reactor.Example/ExamplePlugin.cs
@@ -5,8 +5,7 @@ using Il2CppInterop.Runtime.Attributes;
 using Reactor.Networking;
 using Reactor.Networking.Attributes;
 using Reactor.Networking.Rpc;
-using Reactor.Plugins;
-using Reactor.Plugins.Attributes;
+using Reactor.Utilities;
 using Reactor.Utilities.Attributes;
 using Reactor.Utilities.ImGui;
 using UnityEngine;
