+++ Reactor.Example/ExampleRpc.cs
@@ -1,5 +1,4 @@
 using Hazel;
-using Reactor.Networking;
 using Reactor.Networking.Attributes;
 using Reactor.Networking.Rpc;
