+++ Reactor/Networking/Rpc/CustomRpcManager.cs
@@ -1,6 +1,12 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
+using System.Reflection;
+using HarmonyLib;
+using Hazel;
 using Il2CppInterop.Generator.Extensions;
+using InnerNet;
+using Reactor.Utilities;
 
 namespace Reactor.Networking.Rpc;
 
@@ -17,8 +23,8 @@ public class CustomRpcManager
     {
         customRpc.Manager = this;
         _list.Add(customRpc);
-        _map.GetOrCreate(customRpc.InnerNetObjectType, _ => new Dictionary<Mod, Dictionary<uint, UnsafeCustomRpc>>())
-            .GetOrCreate(customRpc.Mod, _ => new Dictionary<uint, UnsafeCustomRpc>())
+        _map.GetOrCreate(customRpc.InnerNetObjectType, static _ => new Dictionary<Mod, Dictionary<uint, UnsafeCustomRpc>>())
+            .GetOrCreate(customRpc.Mod, static _ => new Dictionary<uint, UnsafeCustomRpc>())
             .Add(customRpc.Id, customRpc);
 
         if (customRpc.IsSingleton)
@@ -27,5 +33,41 @@ public class CustomRpcManager
         }
 
         return customRpc;
+    }
+
+    public static bool HandleRpc(InnerNetObject innerNetObject, byte callId, MessageReader reader)
+    {
+        if (callId == CallId)
+        {
+            var manager = PluginSingleton<ReactorPlugin>.Instance.CustomRpcManager;
+            var customRpcs = manager._map[innerNetObject.GetType()];
+
+            var mod = reader.ReadMod();
+            var id = reader.ReadPackedUInt32();
+
+            var customRpc = customRpcs[mod][id];
+
+            customRpc.UnsafeHandle(innerNetObject, customRpc.UnsafeRead(reader.ReadMessage()));
+
+            return true;
+        }
+
+        return false;
+    }
+
+    [HarmonyPatch]
+    internal static class HandleRpcPatch
+    {
+        private static List<Type> InnerNetObjectTypes { get; } = typeof(InnerNetObject).Assembly.GetTypes().Where(x => x.IsSubclassOf(typeof(InnerNetObject)) && x != typeof(LobbyBehaviour)).ToList();
+
+        public static IEnumerable<MethodBase> TargetMethods()
+        {
+            return InnerNetObjectTypes.Select(x => x.GetMethod(nameof(InnerNetObject.HandleRpc), AccessTools.allDeclared)).Where(m => m != null)!;
+        }
+
+        public static bool Prefix(InnerNetObject __instance, [HarmonyArgument(0)] byte callId, [HarmonyArgument(1)] MessageReader reader)
+        {
+            return !HandleRpc(__instance, callId, reader);
+        }
     }
 }
