+++ Reactor/Patches/Fixes/CursorPosPatch.cs
@@ -8,12 +8,10 @@ namespace Reactor.Patches.Fixes;
 /// <summary>
 /// "Fixes" an issue where empty TextBoxes have wrong cursor positions
 /// </summary>
-[HarmonyPatch]
+[HarmonyPatch(typeof(TextMeshProExtensions), nameof(TextMeshProExtensions.CursorPos))]
 internal static class CursorPosPatch
 {
-    [HarmonyPatch(typeof(TextMeshProExtensions), nameof(TextMeshProExtensions.CursorPos))]
-    [HarmonyPrefix]
-    public static bool FixCursorPosPatch(TextMeshPro self, ref Vector2 __result)
+    public static bool Prefix(TextMeshPro self, ref Vector2 __result)
     {
         if (self.textInfo == null || self.textInfo.lineCount == 0 || self.textInfo.lineInfo[0].characterCount <= 0)
         {
