+++ Reactor/Patches/Miscellaneous/CustomServersPatch.cs
@@ -1,6 +1,6 @@
 using HarmonyLib;
 
-namespace Reactor.Regions.Patches;
+namespace Reactor.Patches.Miscellaneous;
 
 [HarmonyPatch]
 internal static class CustomServersPatch
@@ -8,7 +8,7 @@ internal static class CustomServersPatch
     [HarmonyPatch(typeof(AuthManager._CoConnect_d__4), nameof(AuthManager._CoConnect_d__4.MoveNext))]
     [HarmonyPatch(typeof(AuthManager._CoWaitForNonce_d__6), nameof(AuthManager._CoWaitForNonce_d__6.MoveNext))]
     [HarmonyPrefix]
-    public static bool DisableAuthServerPatch(out bool __result)
+    public static bool DisableAuthServer(out bool __result)
     {
         __result = false;
         return false;
@@ -16,7 +16,7 @@ internal static class CustomServersPatch
 
     [HarmonyPatch(typeof(AmongUsClient._CoJoinOnlinePublicGame_d__1), nameof(AmongUsClient._CoJoinOnlinePublicGame_d__1.MoveNext))]
     [HarmonyPrefix]
-    public static void EnableUdpMatchmakingPatch(AmongUsClient._CoJoinOnlinePublicGame_d__1 __instance)
+    public static void EnableUdpMatchmaking(AmongUsClient._CoJoinOnlinePublicGame_d__1 __instance)
     {
         // Skip to state 1 which just calls CoJoinOnlineGameDirect
         if (__instance.__1__state == 0 && !ServerManager.Instance.IsHttp)
