+++ Reactor/Patches/Miscellaneous/SplashSkipPatch.cs
@@ -1,23 +1,24 @@
 using System.Reflection;
 using HarmonyLib;
 
-namespace Reactor.Patches.QOL;
+namespace Reactor.Patches.Miscellaneous;
 
+[HarmonyPatch]
 internal static class SplashSkipPatch
 {
     private static readonly PropertyInfo? _localUserIdProperty = typeof(EpicManager).GetProperty("localUserId", BindingFlags.Static | BindingFlags.Public);
     private static bool _loginFinished;
-    
+
     [HarmonyPatch(typeof(SplashManager), nameof(SplashManager.Start))]
     [HarmonyPrefix]
-    private static void SplashSkip(SplashManager __instance)
+    private static void RemoveMinimumWait(SplashManager __instance)
     {
         __instance.minimumSecondsBeforeSceneChange = 0;
     }
 
     [HarmonyPatch(typeof(SplashManager), nameof(SplashManager.Update))]
     [HarmonyPrefix]
-    private static bool WaitForEpicAuth(SplashManager __instance)
+    private static bool WaitForLogin(SplashManager __instance)
     {
         if (__instance.startedSceneLoad) return true;
 
@@ -32,7 +33,7 @@ internal static class SplashSkipPatch
     // EpicManager calls SaveManager.LoadPlayerPrefs(true) both on successful and unsuccessful EOS login
     [HarmonyPatch(typeof(SaveManager), nameof(SaveManager.LoadPlayerPrefs))]
     [HarmonyPostfix]
-    private static void LoadPlayerPrefsPatch(bool overrideLoad)
+    private static void WaitForEpicAuth(bool overrideLoad)
     {
         if (overrideLoad) _loginFinished = true;
     }
