+++ Reactor/ReactorPlugin.cs
@@ -1,4 +1,4 @@
-﻿global using static Reactor.Plugins.Logger<Reactor.ReactorPlugin>;
+﻿global using static Reactor.Utilities.Logger<Reactor.ReactorPlugin>;
 using System;
 using System.Linq;
 using BepInEx;
@@ -9,11 +9,8 @@ using Reactor.Networking;
 using Il2CppInterop.Runtime.Attributes;
 using Reactor.Networking.Attributes;
 using Reactor.Networking.Rpc;
-using Reactor.Patches.QOL;
-using Reactor.Patches.Reactor;
-using Reactor.Plugins;
-using Reactor.Regions;
-using Reactor.Resources;
+using Reactor.Patches;
+using Reactor.Patches.Miscellaneous;
 using Reactor.Utilities;
 using Reactor.Utilities.Attributes;
 using UnityEngine;
